+++ src/JSONSchema/index.ts
@@ -1,6 +1,2 @@
 export * from './types'
-export {
-  useObjectFromForm,
-  concatFormPath,
-  useObjectFromPath,
-} from './path-handler'
+export * from './path-handler'
